+++ src/constants/shipComponentTypes.js
@@ -1,0 +1,12 @@
+
+import toMap from 'to-map'
+
+// Immutable key-value map of
+// Use a flat map, makes life a little easier
+const SC_TYPES = toMap({
+    HULL: 'sc:hull',
+    THRUSTER: 'sc:thruster'
+})
+
+
+export default SC_TYPES